renamed person segmentation model to bodyPix
