+++ body-pix/colored-parts.gif
