+++ body-pix/demos/.babelrc
