+++ body-pix/demos/index.html
@@ -2,7 +2,7 @@
 <html>
 
 <head>
-    <title>Person Segmentation - With a Webcam Demo</title>
+    <title>BodyPix - With a Webcam Demo</title>
     <style>
     .footer {
         position: fixed;
@@ -47,7 +47,7 @@
     <div class="footer">
         <div class="footer-text">
             <p>
-              The Person Segmentation model can estimate which pixels in an image are part of a person, and which pixels
+              The BodyPix model can estimate which pixels in an image are part of a person, and which pixels
               are part of each of 24 body parts.  It works on a single person, and such <strong>works best</strong> when <strong>one person is present</strong> in an image.
                 <br>
                 <br> The <strong>output stride</strong> and <strong>model (indicated by mobileNetArchitecture)</strong> have the largest effects on accuracy/speed. A <i>higher</i> output stride results in lower accuracy but higher speed. A <i>larger</i> model, indicated by the <i>mobileNetArchitecture</i> dropdown, results in higher accuracy but lower speed.
