+++ body-pix/demos/part_color_scales.js
