+++ body-pix/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@tensorflow-models/person-segmentation",
+  "name": "@tensorflow-models/body-pix",
   "version": "0.0.4",
-  "description": "Pretrained Person Segmentation model in tensorflow.js",
+  "description": "Pretrained BodyPix model in tensorflow.js",
   "main": "dist/index.js",
-  "jsnext:main": "dist/person-segmentation.esm.js",
-  "module": "dist/person-segmentation.esm.js",
-  "unpkg": "dist/person-segmentation.min.js",
-  "jsdelivr": "dist/person-segmentation.min.js",
+  "jsnext:main": "dist/body-pix.esm.js",
+  "module": "dist/body-pix.esm.js",
+  "unpkg": "dist/body-pix.min.js",
+  "jsdelivr": "dist/body-pix.min.js",
   "types": "dist/index.d.ts",
   "repository": {
     "type": "git",
