+++ body-pix/src/decode_part_map.ts
