+++ body-pix/src/index.ts
@@ -15,9 +15,9 @@
  * =============================================================================
  */
 
+export {BodyPix, load} from './body_pix_model';
 export {Checkpoint, checkpoints} from './checkpoints';
 export {decodePartSegmentation, toMask} from './decode_part_map';
 export {drawBodyMaskOnCanvas, drawBodySegmentsOnCanvas, drawBokehEffectOnCanvas} from './output_rendering_util';
 export {partChannels} from './part_channels';
-export {load, PersonSegmentation} from './person_segmentation_model';
 export {resizeAndPadTo, scaleAndCropToInputTensorShape} from './util';
