+++ body-pix/src/types.ts
@@ -1,4 +1,4 @@
 import * as tf from '@tensorflow/tfjs';
 
-export type PersonSegmentationInput =
+export type BodyPixInput =
     ImageData|HTMLImageElement|HTMLCanvasElement|HTMLVideoElement|tf.Tensor3D;
